add event-emitter
